pyproject.toml
[tool.poetry]
name = "djackets_django"
version = "0.1.0"
description = ""
authors = ["Moritz Gnisia <moritzgnisia@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.1"
djangorestframework = "^3.13.1"
Markdown = "^3.3.6"
django-filter = "^21.1"
django-cors-headers = "^3.11.0"
djoser = "^2.1.0"
Pillow = "^9.0.0"
psycopg2-binary = "^2.9.3"
django-environ = "^0.8.1"
django-extensions = "^3.1.5"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
